refactor(framework): the run-checkout resolution has one home
resolveRunCheckout (daemon.ts) and the tail of resolveRunPath
(dashboard-rpc/context.ts) were the same four steps written twice: safe-id
guard, live metas, worktree directory probe, project-root fallback. The
daemon copy even said 'same resolution the read and control RPCs use' --
a pointer where a shared function should be -- and the #766 subtlety that
makes the directory probe correctness-critical (a Telefunc Channel resolves
its path once and never self-corrects) was documented on only one of them.

It now lives in store/run-checkout.ts beside the inputs it reads
(readLiveMetas, worktreePath); the daemon imports it and resolveRunPath
adds only the project-id lookup.
