+++ packages/framework/src/store/index.ts
@@ -26,6 +26,7 @@ export {
   type RunStatus,
   type OpenStoreOptions,
 } from './run-store.js'
+export { resolveRunCheckout } from './run-checkout.js'
 export {
   addWorktree,
   attachWorktree,
